requirements.txt
langchain
chromadb
streamlit
beautifulsoup4
requests
python-dotenv
groq
schedule